refactor/zustand/my-funnel
