+++ Loopy-fe/src/hooks/Funnel/useMyPageFunnelStore.ts
@@ -1,0 +1,15 @@
+import { create } from "zustand";
+import type { MyPageSteps } from "../../types/mySteps";
+import type { useFunnel } from "@use-funnel/browser";
+
+type FunnelInstance = ReturnType<typeof useFunnel<MyPageSteps>>;
+
+interface MyPageFunnelStore {
+  funnel: FunnelInstance | null;
+  setFunnel: (funnel: FunnelInstance) => void;
+}
+
+export const useMyPageFunnelStore = create<MyPageFunnelStore>((set) => ({
+  funnel: null,
+  setFunnel: (funnel) => set({ funnel }),
+}));
